serve.rs: Create the DB pool ourselves so we can tweak the settings. Fixes weird 5xx errors I've been seeing.
Specifically the fix is the "synchronous" setting of sqlite. The default is
"Full" which does `fsync()` very very often (such that all data is synced before
it tells you things are good). "Normal" does it _way_ less and won't corrupt the
DB[1] (though it might lose data in the RAM WAL on a bad shutdown).

[1] https://www.sqlite.org/pragma.html#pragma_synchronous