syncron.js: In runs view, ask for specific ids to update running jobs (every second).
This took way too long because react is weird. I ended up writing
use_interval_loader() to split off the complexity and now runs_view doesn't
look terrible. I think it may be able to generalize so we could use it in
the jobs view, too (once the server APIs allow it).