migrations/: Add "cooked success column".
Currently sucess is kind of weird and success is defined as:

    status.Exited != null && (status.Exited == 0 || log_len == 0)

This is kind of a hacky heuristic because cron ignores exit status and
only emails based on output and I find it common that cron jobs that
don't print often exit with non-zero status, even when they are
fine. Cron's own man page has this example:

       0 4 8-14 * *    test $(date +\%u) -eq 6 && echo "2nd Saturday"

That returns status 1 on every day it doesn't run.

Perhaps this should be a per job setting? I also have a couple jobs
that never show as failing because they are cavalier about the exit
status in the other direction, such as : `do something ||
cleanup`. It'll appear good but if you check the logs it will have
printed a bunch of errors and failed. Is it reasonable to expect the
user to rewrite their cron job to conform to syncron? (Doing it
"correctly" is tough in the one-liner confines of a crontab). Or
should there be enough configurability to account for different jobs
expressing failure in different ways? (do we need an RE to run against
the log output??)