db.rs: Only store the relative path from the db file in the db so that the --db option works
Before this change, there were 2 possibilities:

  1. --db=../some-relative/path/to/db
  2. --db=/some/absolute/path/to/db

With #1 then your cwd matters: even if you point to the db correctly, the
relative path was stored in the db, so it only works if you run from the
same cwd on subsequent invocations. For both of them it's annoying to move
the db.

Now, the --db points to the base (from the cwd, or absolute), and the db
holds the path from there. That makes the db directory moveable and the cwd
doesn't matter between invocations (as long as --db points relatively from
the cwd on a particular invocation, obviously).