syncron.rs: Add a weird syntax for SYNCRON_NAME to specify the id, instead of using SYNCRON_JOB_ID.
The syntax: "@job-id Job Name" (`@` signals it, ` ` divides it from the name).

I manually renamed a job and when I was adding SYNCRON_JOB_ID to puppet I
realized that I needed to specify it on _every_ job in the crontab that
puppet controlled, otherwise it might "fall through" and get used by another
syncron invocation later in the crontab.

It could be argued that this is a puppet issue, but I realized that even if
I were manually configuring everything I'd have to be vigilant about
unsetting the env after the cron job line. Or I'd have to make the decision
to always set the id and name before any job in the crontab, but that seems
way too much. Also trying to explain this in the documentation would be
annoying.

Putting the ID and the name into the same env variable solves this issue and
makes editing the crontab harder to mess up, which I think more than
balances out the downsides of adding some custom cryptic syntax.

Alternate syntax considered: "@job-id:Job Name". I think it looks a little
better except when the name has a `:` in it, like a lot of my existing names
do:

SYNCRON_NAME="@tmpreaper-var-lib-motion-porch:tmpreaper: /var/lib/motion/Porch/*"

I dunno, I could be convinced.