Add .cargo/config.toml which sets an env var for libsqlite3-sys to make it use the system sqlite
Otherwite it will use its bundled one which is older and doesn't support
jsonb yet. There _is_ one available which is newer but we can't get it since
it's pulled in through rocket_db_pools which hasn't updated to sqlx-0.8 yet
(and we can't depend on it directly or it will conflict).