db.rs,serve.rs: Add optional PruneStats arg to get stats from the prune run2
It's optional because it needs to bypass the optimization for jobs that
aren't set up to be pruned. Large amonuts of runs can be slow since we
(currently) have to stat() the log files to figure out their size.

I'm considering caching/storing the log size in the runs table which would
make everything very fast and remove the need for that optimization and then
the stats could be computed every time with no overhead.