db.rs: Cap number of pruned runs we'll return to 1000
I was trying to run prune on my system and it was filling up memory, I
believe this Vec is the issue. I still like the idea when you're pruning and
trying to figure out why something happened, but it's just not super useful
when there are 10,000+ runs that got pruned.