event.rs, db.rs, serve.rs: Add new event system.
The goal is to feed events to the front ends without making them poll for
info. This should give them faster updates when things are happening and not
waste time when nothing is happening.

Events have a "topic" that subscribers can subscribe to. If this sounds like
MQTT, it's because it was the inspiration. When you subscribe you give a
"topic filter" which uses the same semantics as MQTT: `/` is a topic
component separator, `+` matches any component and `#` matches any number of
components (and can only be at the end). This design makes it very simple to
implement but should be powerful enough to cover our needs.

Currently just the db is hooked up and sends events. The serve.rs part of
the equation will come later. I implemented both a websockets and
server-sent events version. The websockets implementation is 134 lines and
the server-sent events version is just 5 lines, so I'm probably going with
that. But I want to finish the JS UI to make sure I'm not missing any
features.