syncron.js: Switch from using polling to using server sent events endpoint at `/events`
Since server-sent events are limited on HTTP/1.1 (they count towards the
maximum connections open to a server), it will unsubscribe whenever the
document.hidden goes false. When the page becomes visible again it will
download a new snapshot and then resume the events. This _should_ be
sufficient, I think. It seems to work pretty well in my limited testing.