db.rs,server.rs: Change the prune() function parameters to make it more configurable.
We used to pass in Some(&mut PruneStats) to mean "always collect the prune
stats" but that made it impossible to say collect prune stats except when
there's no retention settings for the job.

The new function has a `force_stats` bool that means the same thing, but when
it's false it'll still collect PruneStats on jobs that _do_ have retention
settings. It now return Ok(None) when force_stats is false and the job has
no retention settings. In all other (non-error) cases it returns
Ok(Some((stats, pruned))).